requirements.molecule.txt
#ansible

## ref: https://www.jeffgeerling.com/blog/2024/newer-versions-ansible-dont-work-rhel-8
ansible-core==2.16.3
docker
yamllint
ansible-lint

##'ansible-compat<4' is required due to https://github.com/ansible-community/molecule/issues/3903
## if moving to latest ansible-compat  - we get error "AttributeError: 'Runtime' object has no attribute 'exec'"
#ansible-compat<4
ansible-compat

## ref: https://molecule.readthedocs.io/installation/#pip
## ref: https://github.com/ansible-community/molecule-plugins
#molecule[docker]
molecule
molecule-plugins[docker]

pytest-testinfra

## The ipaddr filter requires python's netaddr be installed on the ansible controller
netaddr

## windowsRM
## ref: https://docs.ansible.com/ansible/latest/user_guide/windows_winrm.html
pywinrm>=0.3.0

## auth related
passlib

## library module requirements
ruamel.yaml
